Item allos (item effects) (#1278)
Item effects (via LibAllo)! This pull changes
- consolidates the many `AccountUseXX` and `KamiUseXX` systems into `AccountUseItem` and `KamiUseItem` 
- uses `LibAllo` pattern for item behaviour
- updates notion page for item entries
- updates item deployment
- client api calls

to implement:
- update client display, clean up code
- separate skill reset system from items


@JirAcheron rather large pull (sorry - the files give me no choice). Would appreciate your review specifically on `KamiItemUseSystem` and the new additions in `LibItem`

+++ packages/contracts/src/libraries/utils/LibGetter.sol
@@ -152,7 +152,7 @@ library LibGetter {
     IUintComp components,
     string memory entityType,
     uint256 id
-  ) internal view returns (uint256) {
+  ) internal view returns (uint32) {
     if (entityType.eq("KAMI")) return LibKami.getStateIndex(components, id);
     else revert("LibGetter: invalid entity state type");
   }
